Reorganized code to allow for mutual defs
Administrative syntax is being placed under the Insn type for now (this
may be split later). This means that we need to bring in run time values
such as closures, exports, etc into the Insn type, which involved some
reorganization of the file to put everything under a `mutual` block.