HEAVY REFACTOR
I refactored the Stack type to reflect that we do not need to keep
AdminInstrs on the stack. This means that Stack is now just an alias for
a List of Vals. Val has moved to the Instr module to make this change
possible.

This means that StackEntries are no longer needed, and neither is the
Constant type. Now there is just a Const instruction that takes as a
parameter a Val. The stack is significantly simplified.

The two tests currently pass but there are certainly bugs introduced
here. With this in mind we should plan on getting some more tests
running.